refactor: :art: seperates different agents strategies
